Move BuildTreeLocations to explicitly internal package
