+++ platforms/core-runtime/build-discovery/src/main/java/org/gradle/internal/initialization/BuildTreeLocations.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package org.gradle.initialization.layout;
+package org.gradle.internal.initialization;
 
+import org.gradle.initialization.layout.BuildLayout;
 import org.gradle.internal.service.scopes.Scope;
 import org.gradle.internal.service.scopes.ServiceScope;
